[cloud-api] Improve logging
Summary:
This commit makes it so we log every request we serve with the path requested,
response code, and the time it takes to serve the request, using the same key
names as we use in the Python API for consistency. It also adds the public ID
of the billing identity to all logs for authorized accounts and turns down the
level of some debug logging that isn't useful in production.

Future diffs coming which consolidate error logging on the request log which
we've found super useful, and also improving local development logging (I'm
gonna fork bunyan-prettystream as it hardcodes which JSON elements it prints).

Test Plan:
need to test on staging - verified basic logging locally but hard
           to test authenticated requests locally (b/c billing)

Reviewers: evan, khamidou, juan

Reviewed By: juan

Differential Revision: https://phab.nylas.com/D3832