[cloud-core] Detect redis disconnections in the PubsubConnector
Summary:
I'm debugging an issue with send-reminders where the metadata deltas
aren't published after the reminder is processed. The cloud-worker
logs show `AbortError: PUBLISH can't be processed. The connection
is already closed`. This error appears on staging, but not locally.
This diff 1) adds a log statement so I can see when the connection
is actually closed and better determine whether this is bug in the
send-reminders code and 2) resets the cached client to null so that
the PubsubConnector will attempt to reconnect for the next delta

Test Plan: staging

Reviewers: juan, evan

Reviewed By: evan

Differential Revision: https://phab.nylas.com/D3995