[send-later][cloud-workers] Fix S3 attachment upload
Summary:
I spent a long time looking for this bug. The problem boils down to our MIME generation library not accepting run of the mill buffer or streams made out of buffers. Amusingly, the author of this library also decided it to make it semi-private; it's on npm but doesn't have any github sources.

To work around this issue I decided to simply write the contents to a temporary file on the workers machine, then read from this file. Works great, even though it's very hackish.

Test Plan: Tested manually.

Reviewers: juan, halla, evan

Reviewed By: halla, evan

Differential Revision: https://phab.nylas.com/D4241